fix: Refactor backend, add sample args
