+++ py/torch_tensorrt/dynamo/__init__.py
@@ -1,0 +1,2 @@
+from _compiler import compile
+from backends import create_backend
