+++ py/torch_tensorrt/dynamo/lowering/__init__.py
@@ -1,0 +1,2 @@
+from torch_tensorrt.dynamo.lowering._decompositions import get_decompositions
+from torch_tensorrt.dynamo.lowering._partition import partition
